export location logic parsing from GenbankSequenceParser to InsdcParser
